Changed Bot.py to main.py so it works on replit and added random join.
