+++ pyproject.toml
@@ -1,0 +1,20 @@
+[tool.poetry]
+name = "repl_nix_Herif-Bot"
+version = "0.1.0"
+description = ""
+authors = ["Replit user <<>>"]
+
+[tool.poetry.dependencies]
+python = "^3.8"
+yt-dlp = "^2022.11.11"
+discord = "^2.1.0"
+Token = "^1.8.6"
+youtube_dl = "^2021.12.17"
+PyNaCl = "^1.5.0"
+Flask = "^2.2.2"
+
+[tool.poetry.dev-dependencies]
+
+[build-system]
+requires = ["poetry-core>=1.0.0"]
+build-backend = "poetry.core.masonry.api"
